[ALPS02673393] kernel: kernel-3.18 porting to kernel-4.4
[Detail]
kernel-3.18 patches need to be ported to kernel-4.4
while kernel-4.4 does not have.

[Solution]
kernel-3.18 patches need to be ported to kernel-4.4
while kernel-4.4 does not have.

MTK-Commit-Id: 526b4e96cf1906d40d8f60a6fa521a8e11c69d59

CR-Id: ALPS02673393
Feature: Others

+++ net/netlink/af_netlink.c
@@ -3174,10 +3174,10 @@ static int netlink_seq_show(struct seq_file *seq, void *v)
 		struct sock *s = v;
 		struct netlink_sock *nlk = nlk_sk(s);
 
-		seq_printf(seq, "%pK %-3d %-6u %08x %-8d %-8d %d %-8d %-8d %-8lu\n",
+		seq_printf(seq, "%pK %-3d %-6d %08x %-8d %-8d %d %-8d %-8d %-8lu\n",
 			   s,
 			   s->sk_protocol,
-			   nlk->portid,
+			   (int)nlk->portid,
 			   nlk->groups ? (u32)nlk->groups[0] : 0,
 			   sk_rmem_alloc_get(s),
 			   sk_wmem_alloc_get(s),
